example/requirements.txt
flask-oauthlib
flask-sqlalchemy
flask-wtf
